fix(ui): events should not hang on loading
Fixes https://discord.com/channels/317348946144002050/619437014001123338/790299708152872982
